watermark_remover/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
pymupdf>=1.23.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.20.0
